task2/requirements.txt
torch>=1.7.0
torchvision>=0.8.0
mmcv>=2.0.0
mmdet>=3.0.0
numpy>=1.18.5
opencv-python>=4.5.0
Pillow>=8.0.0
matplotlib>=3.3.0
tensorboard>=2.6.0